requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
nfl_data_py